Enhancements to expired/cancelled holds reporting
This patch amends the expired holds report to handle the new expiration-date
feature, and creates a separate report for cancelled holds.  The report is
also now filterable on branch.

One SQL change is necessary with this patch:

ALTER TABLE old_reserves DROP FOREIGN KEY `old_reserves_ibfk_3`;

This needs to be done because when someone cancels a hold on a "next
vailable" item, an item number would not be assigned (itemnumber=NULL).
The foreign key constraint would then balk at this because there's not a
NULL itemnumber in the items table.