Bug 3665 Changes to staff advanced search
This is to make the same changes as in Bug 3663 (OPAC advanced search) to the
staff advanced search.

First change is to copy the Search button from the bottom of the screen to up
under the Keyword etc. search boxes, so someone doesn't have to scroll down for
it.  Either button will work.

Second change is to reorder the entries in the search box pulldowns.  I've
rearranged the list so that Keyword, Title, Subject, Author are first in the
list, as I did in the OPAC advanced search.  I moved Barcode and Standard
Number to next in the list to make them more visible to staff, and modified the
description of Standard Number to Standard (biblio) Number because it's not
clear that it means the biblio number.  The rest of the existing search options
follow.