[1/2] Bug 5366 - Standing enh req for code changes enabling database agnosticism
MariaDB apparently will not fudge when one attempts to insert a null value into a
NOT NULL field with no default defined. This probably should have been fixed in
any case, but must be fixed if Koha is to run over MariaDB.

This patch simple supplies a single blank space as the value passed in on the
initial insert.