[3/3] Bug 5366 - Standing enh req for code changes enabling database agnosticism
These issues actually may be between DBD::MySQL and MariaDB rather than between
Koha and MariaDB. However, there is a problem with passing an undef in as a parameter
for a field defaulting to NULL. This path fixes import_borrowers.pl to default those
fields to 0.

This patch also adds code to cause the sample CSV to contain these defaults in the
proper fields of the first row of data.