#5574: follow-up
Owen noticed the batchedit was not working. This was because we made a patch months ago to change the GetBiblio API
(the 1st returned arg is useless)
This patch is not in main trunk, resulting in the error owen has noticed.

This patch goes back to official API (even if it is clumsy : the patch we did does not apply anymore)