Bug 6701 - login timeout is in seconds
I checked in C4/Auth and we are using perl time() for comparison with this value.

I also added description of specifying timeout in days which is available in code.

Robin Sheat suggested to document just lowercase d for days option, to reduce
confusion although uppercase D works as days specifier as well.