Bug 10451 - Firefox prints only first page of table with shrink to fit option
According to Firefox bug tracker, versions up to 21 doesn't print
tables correctly when using shrink to fit option. See:

https://bugzilla.mozilla.org/show_bug.cgi?id=684622

However, even with v21 or newer, we still need to add css to
print media which instruct it not to break table tag to print
all pages of table. This has a side-effect that table starts
always on next page, wasting some space on first printed page
(but only in Firefox since CSS selector is specific to it).