Bug 12831 - local only logins should work with LDAP
This fixes regression introducted in Bug 8148.

If ldap is enabled, and using auth_by_bind, one will not be able to login
using any local accounts which do not appear in the ldap directory.

This is problematic because a number of koha system accounts are unlikely
to have a parallel in the active directory.. for instance SIP Service User
accounts.